examples/serverless/requirements.txt
-i https://pypi.org/simple
certifi==2023.11.17; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
idna==3.6; python_version >= '3.5'
packaging==23.2; python_version >= '3.7'
pillow==10.2.0; python_version >= '3.8'
pytesseract==0.3.10; python_version >= '3.7'
requests==2.31.0; python_version >= '3.7'
urllib3==1.26.6; python_version >= '3.8'